+++ apps/ui-dev/component/navigations/treeview/OptionsView.jsx
@@ -1,8 +1,13 @@
+import ChevronRightFilled from "@elf-framework/icon/ChevronRightFilled";
+// import DescriptionFilled from "@elf-framework/icon/DescriptionFilled";
+// import FolderFilled from "@elf-framework/icon/FolderFilled";
 import { useState } from "@elf-framework/sapa";
-import { Grid, VBox, Slider } from "@elf-framework/ui";
+import { Grid, VBox, TreeViewProvider, TreeView } from "@elf-framework/ui";
+
+import treeData from "./data/base.json";
 
 export function OptionsView() {
-  const [value, setValue] = useState(0);
+  const [provider] = useState(new TreeViewProvider(treeData));
   return (
     <Grid columns={["50%", "50%"]} gap={30}>
       <VBox
@@ -13,23 +18,40 @@ export function OptionsView() {
           fontSize: 13,
         }}
       >
-        <div style={{ display: "flex", flexDirection: "column", padding: 20 }}>
-          <Slider
-            label={"Contrast"}
-            value={value}
-            max={10000}
-            step={20}
-            onInput={(v) => setValue(v)}
-          />
+        <div>
+          <div
+            style={{
+              position: "relative",
+              width: 300,
+              height: 300,
+              margin: "0 auto",
+              flex: "none",
+              borderLeft: "1px solid var(--color-gray-4)",
+              borderRight: "1px solid var(--color-gray-4)",
+            }}
+          >
+            <TreeView
+              // draggable={true}
+              // selectionStyle="checkbox"
+              // renderContext={(item) => {
+              //   return item.data.children?.length ? (
+              //     <FolderFilled />
+              //   ) : (
+              //     <DescriptionFilled />
+              //   );
+              // }}
+              renderArrow={() => {
+                return <ChevronRightFilled />;
+              }}
+              items={provider.items}
+            />
+          </div>
         </div>
       </VBox>
       <div>
         <p>
-          Sliders should always have a label. In rare cases where context is
-          sufficient and an accessibility expert has reviewed the design, the
-          label could be undefined. These sliders should still include an
-          aria-label in HTML (depending on the context, “aria-label” or
-          “aria-labelledby”).
+          A tree view is text-only by default. This option is best used when a
+          hierarchy consists of all of the same content type.
         </p>
       </div>
     </Grid>