+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @elf-framework/design-tokens
 
+## 2.1.39
+
+### Patch Changes
+
+- support ghost, tooltip feature for treeview
+  support tooltip potal
+  add progress circle component
+
 ## 2.1.38
 
 ### Patch Changes