+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @elf-framework/ui
 
+## 2.1.39
+
+### Patch Changes
+
+- support ghost, tooltip feature for treeview
+  support tooltip potal
+  add progress circle component
+- Updated dependencies []:
+  - @elf-framework/design-system@2.1.39
+
 ## 2.1.38
 
 ### Patch Changes